Refactor Maven HTTP cache code to separate module
